fix: cast concept_id to int32 in cast_as_concept_id (issue #7)
Return int32(string_value) instead of the raw string, consistent with
codemap_xwalk_concept_id/codemap_xwalk_source_concept_id and the int32
schema defined in domain_dataframe_column_types.py.

+++ src/ccda_to_omop/value_transformations.py
@@ -99,7 +99,7 @@ def cast_as_concept_id(args_dict):
     string_value = args_dict['input']
     type_value = args_dict['type']
     if type_value == 'CD' or type_value == 'CE':
-        return string_value
+        return int32(string_value)
     else:
         return None
